refactor BlastButton from action-buttons into own file
