+++ src/tools/blast/components/results/BlastResultsButtons.tsx
@@ -1,11 +1,10 @@
 import React, { FC, lazy, useState, Suspense } from 'react';
-import { v1 } from 'uuid';
 import { useHistory } from 'react-router-dom';
-import { useDispatch } from 'react-redux';
 import { sleep } from 'timing-functions';
 import { DownloadIcon, BasketIcon, ReSubmitIcon } from 'franklin-sites';
 
 import SidePanel from '../../../../shared/components/layouts/SidePanel';
+import BlastButton from '../../../../shared/components/action-buttons/Blast';
 
 import { serverParametersToFormParameters } from '../../adapters/BlastParametersAdapter';
 
@@ -16,14 +15,7 @@ import uniProtKBApiUrls, {
 
 import fetchData from '../../../../shared/utils/fetchData';
 
-import { addMessage } from '../../../../messages/state/messagesActions';
-
 import { PublicServerParameters } from '../../types/blastServerParameters';
-import { UniProtkbAPIModel } from '../../../../uniprotkb/adapters/uniProtkbConverter';
-import {
-  MessageFormat,
-  MessageLevel,
-} from '../../../../messages/types/messagesTypes';
 import { Suggestions } from '../../../../uniprotkb/components/query-builder/AutocompleteWrapper';
 
 type ResubmitButtonProps = {
@@ -88,90 +80,6 @@ const ResubmitButton: FC<ResubmitButtonProps> = ({ inputParamsData }) => {
     >
       <ReSubmitIcon />
       Resubmit
-    </button>
-  );
-};
-
-type BlastButtonProps = {
-  selectedEntries: string[];
-};
-
-const BlastButton: FC<BlastButtonProps> = ({ selectedEntries }) => {
-  const history = useHistory();
-  const dispatch = useDispatch();
-
-  const [disabled, setDisabled] = useState(false);
-
-  const handleClick = async () => {
-    if (selectedEntries.length !== 1) {
-      return;
-    }
-
-    setDisabled(true);
-
-    const [accession] = selectedEntries;
-
-    try {
-      const response = await fetchData<UniProtkbAPIModel>(
-        uniProtKBApiUrls.entry(accession)
-      );
-      const entry = response.data;
-
-      // build FASTA
-      let sequence = entry.sequence.value;
-      const db = entry.entryType.includes('unreviewed') ? 'tr' : 'sp';
-      let optionalProteinName = '';
-      if (entry?.proteinDescription?.recommendedName?.fullName) {
-        optionalProteinName = `${entry.proteinDescription.recommendedName.fullName.value} `;
-      }
-      let optionalOS = '';
-      if (entry?.organism?.scientificName) {
-        optionalOS = `OS=${entry.organism.scientificName} `;
-      }
-      let optionalOX = '';
-      if (entry?.organism?.taxonId) {
-        optionalOX = `OX=${entry.organism.taxonId} `;
-      }
-      let optionalGN = '';
-      if (entry.genes?.[0]?.geneName?.value) {
-        optionalGN = `GN=${entry.genes[0].geneName.value} `;
-      }
-      const pe = entry.proteinExistence[0];
-      let optionalSV = '';
-      if (entry?.entryAudit?.sequenceVersion) {
-        optionalSV = `SV=${entry.entryAudit.sequenceVersion} `;
-      }
-      sequence = `>${db}|${entry.primaryAccession}|${entry.uniProtkbId} ${optionalProteinName}${optionalOS}${optionalOX} ${optionalGN}PE=${pe}${optionalSV}\n${sequence}`;
-
-      history.push(LocationToPath[Location.Blast], {
-        parameters: { sequence },
-      });
-    } catch (err) {
-      setDisabled(false);
-
-      if (!(err instanceof Error)) {
-        return;
-      }
-
-      dispatch(
-        addMessage({
-          id: v1(),
-          content: err.message,
-          format: MessageFormat.POP_UP,
-          level: MessageLevel.FAILURE,
-        })
-      );
-    }
-  };
-
-  return (
-    <button
-      type="button"
-      className="button tertiary"
-      disabled={selectedEntries.length !== 1 || disabled}
-      onClick={handleClick}
-    >
-      Blast
     </button>
   );
 };
